Closes #130: Calculate shipping costs in SDK
The Order::$shippingCosts field will be automatically populated when
calling reserve() on an order.

There is the additional convinience method on the SDK
calculateShippingCosts(array $products) to calculate the shipping costs
for a set of products.

The calculation receives the general shipping costs for a shop
(configured in the social network) and ignore products with free
delivery. The shipping costs are only added once for every shop, no
matter how many products are delivered.